chore(GlobStream): Move tests to new file
- Move all GlobStream tests to their own file
- Remove unnecessary dependencies in package.json and in specs
- Refactor through2 to use miss.through
