+++ package.json
@@ -31,7 +31,6 @@
     "pumpify": "^1.3.5",
     "readable-stream": "^2.1.5",
     "remove-trailing-separator": "^1.0.1",
-    "through2": "^0.6.0",
     "to-absolute-glob": "^2.0.0",
     "unique-stream": "^2.0.2"
   },
